refactor packaging and bind end runtimes with services
