+++ flagd-proxy/pkg/service/handler.go
@@ -1,4 +1,4 @@
-package sync
+package service
 
 import (
 	"context"
@@ -9,8 +9,8 @@ import (
 	syncv12 "buf.build/gen/go/open-feature/flagd/protocolbuffers/go/flagd/sync/v1"
 	syncv1 "buf.build/gen/go/open-feature/flagd/protocolbuffers/go/sync/v1"
 	"github.com/open-feature/flagd/core/pkg/logger"
-	"github.com/open-feature/flagd/core/pkg/subscriptions"
 	"github.com/open-feature/flagd/core/pkg/sync"
+	"github.com/open-feature/flagd/flagd-proxy/pkg/service/subscriptions"
 )
 
 type handler struct {
