+++ flagd-proxy/pkg/service/subscriptions/manager.go
