+++ flagd/pkg/service/flag-evaluation/connect_service.go
@@ -16,11 +16,11 @@ import (
 	"github.com/open-feature/flagd/core/pkg/evaluator"
 	"github.com/open-feature/flagd/core/pkg/logger"
 	"github.com/open-feature/flagd/core/pkg/service"
-	"github.com/open-feature/flagd/core/pkg/service/middleware"
-	corsmw "github.com/open-feature/flagd/core/pkg/service/middleware/cors"
-	h2cmw "github.com/open-feature/flagd/core/pkg/service/middleware/h2c"
-	metricsmw "github.com/open-feature/flagd/core/pkg/service/middleware/metrics"
 	"github.com/open-feature/flagd/core/pkg/telemetry"
+	"github.com/open-feature/flagd/flagd/pkg/service/middleware"
+	corsmw "github.com/open-feature/flagd/flagd/pkg/service/middleware/cors"
+	h2cmw "github.com/open-feature/flagd/flagd/pkg/service/middleware/h2c"
+	metricsmw "github.com/open-feature/flagd/flagd/pkg/service/middleware/metrics"
 	"github.com/prometheus/client_golang/prometheus/promhttp"
 	"go.uber.org/zap"
 	"golang.org/x/net/http2"
