+++ flagd/pkg/service/flag-evaluation/connect_service_test.go
@@ -16,8 +16,8 @@ import (
 	"github.com/open-feature/flagd/core/pkg/logger"
 	"github.com/open-feature/flagd/core/pkg/model"
 	iservice "github.com/open-feature/flagd/core/pkg/service"
-	middlewaremock "github.com/open-feature/flagd/core/pkg/service/middleware/mock"
 	"github.com/open-feature/flagd/core/pkg/telemetry"
+	"github.com/open-feature/flagd/flagd/pkg/service/middleware/mock"
 	"github.com/stretchr/testify/require"
 	"go.opentelemetry.io/otel/sdk/metric"
 	"go.opentelemetry.io/otel/sdk/resource"
